[3.0] Tabulky na webu
- Přidány HTML tabulky s pomocí Express.js čerpající živě z Discord serverů a DB
- Aktualizováno dotenv
- Aktualizovány LSSD admin role
- Přidána ochrana pro mazání dat DB z jiného sboru
- Příkazy /kolega a /hledat byly nahrazeny funkčním příkazem /profil
- Přidány vyšší Leadership hodnosti sborů pro přidání do DB
- Upraveny příklady v DB složkách

+++ src/commands/db.js
@@ -1,7 +1,7 @@
-import { ActionRowBuilder, ModalBuilder, SlashCommandBuilder, TextInputBuilder, TextInputStyle } from "discord.js";
+import { ActionRowBuilder, ButtonBuilder, ButtonStyle, ModalBuilder, SlashCommandBuilder, TextInputBuilder, TextInputStyle } from "discord.js";
 import fs from "fs";
 import path from "path";
-import { checkDB } from "../../src/functions/db.js";
+import { checkDB, getDB } from "../../src/functions/db.js";
 
 export const slash = new SlashCommandBuilder()
     .setName("db")
@@ -30,11 +30,20 @@ export default async function run(bot, i) {
     let passed = false;
     await i.guild.fetch();
     const admin = await i.member;
-    if (admin.roles.cache.has("1145344761402765343")) passed = true; //Staff team Refresh
-    if (admin.roles.cache.has("1139266408681844887")) passed = true; //.
-    if (admin.id === "607915400604286997") passed = true; //Samus
-    if (admin.id === "436180906533715969") passed = true; //Mičut
-    if (admin.id === "411436203330502658") passed = true; //PetyXbron
+    if (admin.id === "411436203330502658") passed = true; //PetyXbron / b1ngo
+    if (bot.LEA.g.SAHP.includes(i.guild.id) && !passed) {
+        if (admin.roles.cache.has("1145344761402765343")) passed = true; //Staff team Refresh
+        if (admin.roles.cache.has("1139266408681844887")) passed = true; //.
+        if (admin.id === "607915400604286997") passed = true; //Samus
+        if (admin.id === "436180906533715969") passed = true; //Mičut
+    } else if (bot.LEA.g.LSSD.includes(i.guild.id) && !passed) {
+        if (admin.roles.cache.has("1167182546904293481")) passed = true; //Staff team Refresh
+        if (admin.roles.cache.has("1167182546904293482")) passed = true; //*
+        if (admin.roles.cache.has("1190825815596875829")) passed = true; //.
+        if (admin.id === "798644986215661589") passed = true; //Smouky
+        if (admin.id === "829978476701941781") passed = true; //Frexik
+    }
+
     if (!passed) return i.reply({ content: "> 🛑 **K tomuhle má přístup jen admin.**", ephemeral: true });
 
     if (choice === "p") {
@@ -92,16 +101,25 @@ export default async function run(bot, i) {
     } else if (choice === "z") {
         if (!(await checkDB(user.id))) return i.reply({ content: "> 🛑 <@" + user.id + "> **není v DB.**", ephemeral: true });
 
-        let log;
-        if (bot.LEA.g.SAHP.includes(i.guild.id)) log = path.resolve("./db/SAHP") + "/" + user.id + ".json";
-        else if (bot.LEA.g.LSSD.includes(i.guild.id)) log = path.resolve("./db/LSSD") + "/" + user.id + ".json";
+        let log, sbor;
+        if (bot.LEA.g.SAHP.includes(i.guild.id)) log = path.resolve("./db/SAHP") + "/" + user.id + ".json", sbor = "SAHP";
+        else if (bot.LEA.g.LSSD.includes(i.guild.id)) log = path.resolve("./db/LSSD") + "/" + user.id + ".json", sbor = "LSSD";
         else return i.reply({ content: "> 🛑 **Tenhle server není uveden a seznamu.**\nKontaktuj majitele (viz. </menu:1170376396678377596>).", ephemeral: true });
+
+        if (!fs.existsSync(log)) {
+            if (bot.LEA.g.SAHP.includes(i.guild.id)) log = path.resolve("./db/LSSD") + "/" + user.id + ".json", sbor = "LSSD";
+            else if (bot.LEA.g.LSSD.includes(i.guild.id)) log = path.resolve("./db/SAHP") + "/" + user.id + ".json", sbor = "SAHP";
+        }
 
         console.log(" < [CMD/DB] >  " + i.member.displayName + ` zobrazil(a) DB záznam ${user.id}.json`);
 
-        i.reply({ files: [log], ephemeral: true });
+        i.reply({ content: `> ✅ **<@${user.id}>, členem \`${sbor}\`**`, files: [log], ephemeral: true });
     } else if (choice === "r") {
         if (!(await checkDB(user.id))) return i.reply({ content: "> 🛑 <@" + user.id + "> **už není v DB.**", ephemeral: true });
+
+        const gotDB = await getDB(user.id);
+        if (!bot.LEA.g[gotDB.guildName].includes(i.guild.id)) return i.reply({ content: `> 🛑 **<@${user.id}> je členem \`${gotDB.guildName}\`!** (Nemůžeš ho povýšit)`, ephemeral: true });
+
         const modal = new ModalBuilder()
             .setCustomId("rankUpModal")
             .setTitle("SAHP | Povýšení");
@@ -160,11 +178,52 @@ export default async function run(bot, i) {
         else if (bot.LEA.g.LSSD.includes(i.guild.id)) loc = path.resolve("./db/LSSD") + "/" + user.id + ".json";
         else return i.reply({ content: "> 🛑 **Tenhle server není uveden a seznamu.**\nKontaktuj majitele (viz. </menu:1170376396678377596>).", ephemeral: true });
 
-        let log = fs.readFileSync(loc, "utf-8");
-        fs.unlinkSync(loc);
+        const admins = [
+            "411436203330502658"/*b1ngo*/, "607915400604286997"/*samus*/, "436180906533715969",/*micut*/
+            "829978476701941781"/*frexikk*/, "798644986215661589"/*smouky*/
+        ];
 
-        console.log(" < [CMD/DB] >  " + i.member.displayName + ` smazal(a) DB záznam ${user.id}.json`);
+        if (!fs.existsSync(loc)) {
+            if (!admins.includes(admin.id)) return i.reply({ content: "> 🛑 **<@" + user.id + "> je v jiném sboru. Nemůžeš ho odebrat!**", ephemeral: true });
 
-        i.reply({ content: `\`\`\`json\n${log}\`\`\`Tenhle záznam (<@${user.id}>) byl vymazán z DB!`, ephemeral: true });
+            const row = new ActionRowBuilder()
+                .addComponents(
+                    new ButtonBuilder()
+                        .setCustomId('confirmOtherSborDelete')
+                        .setLabel('Opravdu smazat')
+                        .setStyle(ButtonStyle.Danger)
+                        .setEmoji('🛑'),
+                );
+            const rpl = await i.reply({ content: "> ⚠️ **<@" + user.id + "> je v DB jiného sboru. Opravdu chceš záznam odebrat?** *(30s na odpověď)*", ephemeral: true, components: [row] });
+
+            const filter = i => i.customId === 'confirmOtherSborDelete' && i.user.id === admin.id;
+
+            const collector = rpl.createMessageComponentCollector({
+                filter, max: 1, time: 30000
+            });
+
+            collector.on('collect', async i => {
+                if (bot.LEA.g.SAHP.includes(i.guild.id)) loc = path.resolve("./db/LSSD") + "/" + user.id + ".json";
+                else if (bot.LEA.g.LSSD.includes(i.guild.id)) loc = path.resolve("./db/SAHP") + "/" + user.id + ".json";
+
+                fs.unlinkSync(loc);
+                console.log(" < [CMD/DB] >  " + i.member.displayName + ` smazal(a) DB záznam ${user.id}.json`);
+
+                return await rpl.edit({ content: `**Tenhle záznam (<@${user.id}>) byl vymazán z DB!**`, files: [loc], components: [] });
+            });
+
+            collector.on('error', async () => {
+
+                return await rpl.edit({ content: "> 🛑 **Čas vypršel. Záznam nebyl smazán.**", components: [] });
+            });
+
+            collector.on('end', async collected => {
+                if (collected.size === 0) return await rpl.edit({ content: "> 🛑 **Čas vypršel. Záznam nebyl smazán.**", components: [] });
+            });
+        } else {
+            i.reply({ content: `**Tenhle záznam (<@${user.id}>) byl vymazán z DB!**`, files: [loc], ephemeral: true });
+            console.log(" < [CMD/DB] >  " + i.member.displayName + ` smazal(a) DB záznam ${user.id}.json`);
+            return fs.unlinkSync(loc);
+        }
     }
 };
